# admin/generated/oas_router_gen.go
// Code generated by ogen, DO NOT EDIT.

package generated

import (
	"net/http"
	"net/url"
	"strings"

	"github.com/ogen-go/ogen/uri"
)

func (s *Server) cutPrefix(path string) (string, bool) {
	prefix := s.cfg.Prefix
	if prefix == "" {
		return path, true
	}
	if !strings.HasPrefix(path, prefix) {
		// Prefix doesn't match.
		return "", false
	}
	// Cut prefix from the path.
	return strings.TrimPrefix(path, prefix), true
}

// ServeHTTP serves http request as defined by OpenAPI v3 specification,
// calling handler that matches the path or returning not found error.
func (s *Server) ServeHTTP(w http.ResponseWriter, r *http.Request) {
	elem := r.URL.Path
	elemIsEscaped := false
	if rawPath := r.URL.RawPath; rawPath != "" {
		if normalized, ok := uri.NormalizeEscapedPath(rawPath); ok {
			elem = normalized
			elemIsEscaped = strings.ContainsRune(elem, '%')
		}
	}

	elem, ok := s.cutPrefix(elem)
	if !ok || len(elem) == 0 {
		s.notFound(w, r)
		return
	}
	args := [1]string{}

	// Static code generated router with unwrapped path search.
	switch {
	default:
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/admin/v1/api/"
			if l := len("/admin/v1/api/"); len(elem) >= l && elem[0:l] == "/admin/v1/api/" {
				elem = elem[l:]
			} else {
				break
			}

			if len(elem) == 0 {
				break
			}
			switch elem[0] {
			case 'a': // Prefix: "auth/"
				if l := len("auth/"); len(elem) >= l && elem[0:l] == "auth/" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					break
				}
				switch elem[0] {
				case 'l': // Prefix: "log"
					if l := len("log"); len(elem) >= l && elem[0:l] == "log" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						break
					}
					switch elem[0] {
					case 'i': // Prefix: "in"
						if l := len("in"); len(elem) >= l && elem[0:l] == "in" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch r.Method {
							case "POST":
								s.handleAdminV1APIAuthLoginPostRequest([0]string{}, elemIsEscaped, w, r)
							default:
								s.notAllowed(w, r, "POST")
							}

							return
						}
					case 'o': // Prefix: "out"
						if l := len("out"); len(elem) >= l && elem[0:l] == "out" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch r.Method {
							case "POST":
								s.handleAdminV1APIAuthLogoutPostRequest([0]string{}, elemIsEscaped, w, r)
							default:
								s.notAllowed(w, r, "POST")
							}

							return
						}
					}
				case 'r': // Prefix: "registration"
					if l := len("registration"); len(elem) >= l && elem[0:l] == "registration" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "POST":
							s.handleAdminV1APIAuthRegistrationPostRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, "POST")
						}

						return
					}
				}
			case 'b': // Prefix: "blog"
				if l := len("blog"); len(elem) >= l && elem[0:l] == "blog" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					// Leaf node.
					switch r.Method {
					case "GET":
						s.handleAdminV1APIBlogGetRequest([0]string{}, elemIsEscaped, w, r)
					case "PATCH":
						s.handleAdminV1APIBlogPatchRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, "GET,PATCH")
					}

					return
				}
			case 'i': // Prefix: "image"
				if l := len("image"); len(elem) >= l && elem[0:l] == "image" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					switch r.Method {
					case "DELETE":
						s.handleAdminV1APIImageDeleteRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, "DELETE")
					}

					return
				}
				switch elem[0] {
				case 's': // Prefix: "s/"
					if l := len("s/"); len(elem) >= l && elem[0:l] == "s/" {
						elem = elem[l:]
					} else {
						break
					}

					// Param: "number"
					// Leaf parameter
					args[0] = elem
					elem = ""

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "GET":
							s.handleAdminV1APIImagesNumberGetRequest([1]string{
								args[0],
							}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, "GET")
						}

						return
					}
				}
			case 'p': // Prefix: "post"
				if l := len("post"); len(elem) >= l && elem[0:l] == "post" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					switch r.Method {
					case "PATCH":
						s.handleAdminV1APIPostPatchRequest([0]string{}, elemIsEscaped, w, r)
					case "POST":
						s.handleAdminV1APIPostPostRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, "PATCH,POST")
					}

					return
				}
				switch elem[0] {
				case '/': // Prefix: "/"
					if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
						elem = elem[l:]
					} else {
						break
					}

					// Param: "postId"
					// Leaf parameter
					args[0] = elem
					elem = ""

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "DELETE":
							s.handleAdminV1APIPostPostIdDeleteRequest([1]string{
								args[0],
							}, elemIsEscaped, w, r)
						case "GET":
							s.handleAdminV1APIPostPostIdGetRequest([1]string{
								args[0],
							}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, "DELETE,GET")
						}

						return
					}
				case 's': // Prefix: "s/"
					if l := len("s/"); len(elem) >= l && elem[0:l] == "s/" {
						elem = elem[l:]
					} else {
						break
					}

					// Param: "number"
					// Leaf parameter
					args[0] = elem
					elem = ""

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "GET":
							s.handleAdminV1APIPostsNumberGetRequest([1]string{
								args[0],
							}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, "GET")
						}

						return
					}
				}
			case 'u': // Prefix: "u"
				if l := len("u"); len(elem) >= l && elem[0:l] == "u" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					break
				}
				switch elem[0] {
				case 'p': // Prefix: "pload"
					if l := len("pload"); len(elem) >= l && elem[0:l] == "pload" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "POST":
							s.handleAdminV1APIUploadPostRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, "POST")
						}

						return
					}
				case 's': // Prefix: "ser"
					if l := len("ser"); len(elem) >= l && elem[0:l] == "ser" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						switch r.Method {
						case "GET":
							s.handleAdminV1APIUserGetRequest([0]string{}, elemIsEscaped, w, r)
						case "PATCH":
							s.handleAdminV1APIUserPatchRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, "GET,PATCH")
						}

						return
					}
					switch elem[0] {
					case '/': // Prefix: "/"
						if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
							elem = elem[l:]
						} else {
							break
						}

						// Param: "id"
						// Leaf parameter
						args[0] = elem
						elem = ""

						if len(elem) == 0 {
							// Leaf node.
							switch r.Method {
							case "GET":
								s.handleAdminV1APIUserIDGetRequest([1]string{
									args[0],
								}, elemIsEscaped, w, r)
							default:
								s.notAllowed(w, r, "GET")
							}

							return
						}
					case 'i': // Prefix: "id"
						if l := len("id"); len(elem) >= l && elem[0:l] == "id" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch r.Method {
							case "GET":
								s.handleAdminV1APIUseridGetRequest([0]string{}, elemIsEscaped, w, r)
							default:
								s.notAllowed(w, r, "GET")
							}

							return
						}
					}
				}
			}
		}
	}
	s.notFound(w, r)
}

// Route is route object.
type Route struct {
	name        string
	summary     string
	operationID string
	pathPattern string
	count       int
	args        [1]string
}

// Name returns ogen operation name.
//
// It is guaranteed to be unique and not empty.
func (r Route) Name() string {
	return r.name
}

// Summary returns OpenAPI summary.
func (r Route) Summary() string {
	return r.summary
}

// OperationID returns OpenAPI operationId.
func (r Route) OperationID() string {
	return r.operationID
}

// PathPattern returns OpenAPI path.
func (r Route) PathPattern() string {
	return r.pathPattern
}

// Args returns parsed arguments.
func (r Route) Args() []string {
	return r.args[:r.count]
}

// FindRoute finds Route for given method and path.
//
// Note: this method does not unescape path or handle reserved characters in path properly. Use FindPath instead.
func (s *Server) FindRoute(method, path string) (Route, bool) {
	return s.FindPath(method, &url.URL{Path: path})
}

// FindPath finds Route for given method and URL.
func (s *Server) FindPath(method string, u *url.URL) (r Route, _ bool) {
	var (
		elem = u.Path
		args = r.args
	)
	if rawPath := u.RawPath; rawPath != "" {
		if normalized, ok := uri.NormalizeEscapedPath(rawPath); ok {
			elem = normalized
		}
		defer func() {
			for i, arg := range r.args[:r.count] {
				if unescaped, err := url.PathUnescape(arg); err == nil {
					r.args[i] = unescaped
				}
			}
		}()
	}

	elem, ok := s.cutPrefix(elem)
	if !ok {
		return r, false
	}

	// Static code generated router with unwrapped path search.
	switch {
	default:
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/admin/v1/api/"
			if l := len("/admin/v1/api/"); len(elem) >= l && elem[0:l] == "/admin/v1/api/" {
				elem = elem[l:]
			} else {
				break
			}

			if len(elem) == 0 {
				break
			}
			switch elem[0] {
			case 'a': // Prefix: "auth/"
				if l := len("auth/"); len(elem) >= l && elem[0:l] == "auth/" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					break
				}
				switch elem[0] {
				case 'l': // Prefix: "log"
					if l := len("log"); len(elem) >= l && elem[0:l] == "log" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						break
					}
					switch elem[0] {
					case 'i': // Prefix: "in"
						if l := len("in"); len(elem) >= l && elem[0:l] == "in" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							switch method {
							case "POST":
								// Leaf: AdminV1APIAuthLoginPost
								r.name = "AdminV1APIAuthLoginPost"
								r.summary = ""
								r.operationID = ""
								r.pathPattern = "/admin/v1/api/auth/login"
								r.args = args
								r.count = 0
								return r, true
							default:
								return
							}
						}
					case 'o': // Prefix: "out"
						if l := len("out"); len(elem) >= l && elem[0:l] == "out" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							switch method {
							case "POST":
								// Leaf: AdminV1APIAuthLogoutPost
								r.name = "AdminV1APIAuthLogoutPost"
								r.summary = ""
								r.operationID = ""
								r.pathPattern = "/admin/v1/api/auth/logout"
								r.args = args
								r.count = 0
								return r, true
							default:
								return
							}
						}
					}
				case 'r': // Prefix: "registration"
					if l := len("registration"); len(elem) >= l && elem[0:l] == "registration" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						switch method {
						case "POST":
							// Leaf: AdminV1APIAuthRegistrationPost
							r.name = "AdminV1APIAuthRegistrationPost"
							r.summary = ""
							r.operationID = ""
							r.pathPattern = "/admin/v1/api/auth/registration"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}
				}
			case 'b': // Prefix: "blog"
				if l := len("blog"); len(elem) >= l && elem[0:l] == "blog" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					switch method {
					case "GET":
						// Leaf: AdminV1APIBlogGet
						r.name = "AdminV1APIBlogGet"
						r.summary = ""
						r.operationID = ""
						r.pathPattern = "/admin/v1/api/blog"
						r.args = args
						r.count = 0
						return r, true
					case "PATCH":
						// Leaf: AdminV1APIBlogPatch
						r.name = "AdminV1APIBlogPatch"
						r.summary = ""
						r.operationID = ""
						r.pathPattern = "/admin/v1/api/blog"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}
			case 'i': // Prefix: "image"
				if l := len("image"); len(elem) >= l && elem[0:l] == "image" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					switch method {
					case "DELETE":
						r.name = "AdminV1APIImageDelete"
						r.summary = ""
						r.operationID = ""
						r.pathPattern = "/admin/v1/api/image"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}
				switch elem[0] {
				case 's': // Prefix: "s/"
					if l := len("s/"); len(elem) >= l && elem[0:l] == "s/" {
						elem = elem[l:]
					} else {
						break
					}

					// Param: "number"
					// Leaf parameter
					args[0] = elem
					elem = ""

					if len(elem) == 0 {
						switch method {
						case "GET":
							// Leaf: AdminV1APIImagesNumberGet
							r.name = "AdminV1APIImagesNumberGet"
							r.summary = ""
							r.operationID = ""
							r.pathPattern = "/admin/v1/api/images/{number}"
							r.args = args
							r.count = 1
							return r, true
						default:
							return
						}
					}
				}
			case 'p': // Prefix: "post"
				if l := len("post"); len(elem) >= l && elem[0:l] == "post" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					switch method {
					case "PATCH":
						r.name = "AdminV1APIPostPatch"
						r.summary = ""
						r.operationID = ""
						r.pathPattern = "/admin/v1/api/post"
						r.args = args
						r.count = 0
						return r, true
					case "POST":
						r.name = "AdminV1APIPostPost"
						r.summary = ""
						r.operationID = ""
						r.pathPattern = "/admin/v1/api/post"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}
				switch elem[0] {
				case '/': // Prefix: "/"
					if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
						elem = elem[l:]
					} else {
						break
					}

					// Param: "postId"
					// Leaf parameter
					args[0] = elem
					elem = ""

					if len(elem) == 0 {
						switch method {
						case "DELETE":
							// Leaf: AdminV1APIPostPostIdDelete
							r.name = "AdminV1APIPostPostIdDelete"
							r.summary = ""
							r.operationID = ""
							r.pathPattern = "/admin/v1/api/post/{postId}"
							r.args = args
							r.count = 1
							return r, true
						case "GET":
							// Leaf: AdminV1APIPostPostIdGet
							r.name = "AdminV1APIPostPostIdGet"
							r.summary = ""
							r.operationID = ""
							r.pathPattern = "/admin/v1/api/post/{postId}"
							r.args = args
							r.count = 1
							return r, true
						default:
							return
						}
					}
				case 's': // Prefix: "s/"
					if l := len("s/"); len(elem) >= l && elem[0:l] == "s/" {
						elem = elem[l:]
					} else {
						break
					}

					// Param: "number"
					// Leaf parameter
					args[0] = elem
					elem = ""

					if len(elem) == 0 {
						switch method {
						case "GET":
							// Leaf: AdminV1APIPostsNumberGet
							r.name = "AdminV1APIPostsNumberGet"
							r.summary = ""
							r.operationID = ""
							r.pathPattern = "/admin/v1/api/posts/{number}"
							r.args = args
							r.count = 1
							return r, true
						default:
							return
						}
					}
				}
			case 'u': // Prefix: "u"
				if l := len("u"); len(elem) >= l && elem[0:l] == "u" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					break
				}
				switch elem[0] {
				case 'p': // Prefix: "pload"
					if l := len("pload"); len(elem) >= l && elem[0:l] == "pload" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						switch method {
						case "POST":
							// Leaf: AdminV1APIUploadPost
							r.name = "AdminV1APIUploadPost"
							r.summary = ""
							r.operationID = ""
							r.pathPattern = "/admin/v1/api/upload"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}
				case 's': // Prefix: "ser"
					if l := len("ser"); len(elem) >= l && elem[0:l] == "ser" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						switch method {
						case "GET":
							r.name = "AdminV1APIUserGet"
							r.summary = ""
							r.operationID = ""
							r.pathPattern = "/admin/v1/api/user"
							r.args = args
							r.count = 0
							return r, true
						case "PATCH":
							r.name = "AdminV1APIUserPatch"
							r.summary = ""
							r.operationID = ""
							r.pathPattern = "/admin/v1/api/user"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}
					switch elem[0] {
					case '/': // Prefix: "/"
						if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
							elem = elem[l:]
						} else {
							break
						}

						// Param: "id"
						// Leaf parameter
						args[0] = elem
						elem = ""

						if len(elem) == 0 {
							switch method {
							case "GET":
								// Leaf: AdminV1APIUserIDGet
								r.name = "AdminV1APIUserIDGet"
								r.summary = ""
								r.operationID = ""
								r.pathPattern = "/admin/v1/api/user/{id}"
								r.args = args
								r.count = 1
								return r, true
							default:
								return
							}
						}
					case 'i': // Prefix: "id"
						if l := len("id"); len(elem) >= l && elem[0:l] == "id" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							switch method {
							case "GET":
								// Leaf: AdminV1APIUseridGet
								r.name = "AdminV1APIUseridGet"
								r.summary = ""
								r.operationID = ""
								r.pathPattern = "/admin/v1/api/userid"
								r.args = args
								r.count = 0
								return r, true
							default:
								return
							}
						}
					}
				}
			}
		}
	}
	return r, false
}